Unbearably lag

+++ apps/gsrast/shaders/splats/fragment.glsl
@@ -23,14 +23,14 @@ vec3 rsIntersection(out vec3 normal) {
     vec3 localRo = (camPos - ellipsoidCenter) * ellipsoidRot;
     vec3 localRd = normalize(normalize(position - camPos) * ellipsoidRot);
 
-    dvec3 oneOver = 1.0 / dvec3(ellipsoidScale);
+    dvec3 oneOver = double(1.0) / dvec3(ellipsoidScale);
 
     dvec3 localRoScl = dvec3(localRo) * oneOver;
     dvec3 localRdScl = dvec3(localRd) * oneOver;
 
     double a = dot(localRdScl, localRdScl);
-    double b = 2.0 * dot(localRoScl, localRdScl);
-    double c = dot(localRoScl, localRoScl) - 1.0; // We have normalized the radius by dividing it by scale
+    double b = double(2.0) * dot(localRoScl, localRdScl);
+    double c = dot(localRoScl, localRoScl) - double(1.0); // We have normalized the radius by dividing it by scale
 
     double discriminant = b * b - 4.0 * a * c;
 
@@ -38,9 +38,9 @@ vec3 rsIntersection(out vec3 normal) {
         return vec3(0.0);
     }
 
-    double t1 = (-b + sqrt(discriminant)) / (2.0 * a);
-    double t2 = (-b - sqrt(discriminant)) / (2.0 * a);
-    float t = float(min(t1, t2));
+    float t1 = float((-b + sqrt(discriminant)) / (2.0 * a));
+    float t2 = float((-b - sqrt(discriminant)) / (2.0 * a));
+    float t = min(t1, t2);
 
     vec3 localPos = localRo + t * localRd;
     vec3 localN = normalize(localPos / ellipsoidScale);
